Use distance only instead of number of points + distance
