+++ core/src/test/java/com/graphhopper/reader/dem/GraphElevationSmoothingTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertEquals;
 /**
  * @author Robin Boldt
  */
-public class RoadAverageElevationInterpolatorTest {
+public class GraphElevationSmoothingTest {
 
     @Test
     public void interpolatesElevationOfPillarNodes() {
@@ -34,7 +34,7 @@ public class RoadAverageElevationInterpolatorTest {
         pl1.add(0, 0, 0);
         pl1.add(0.0005, 0.0005, 100);
         pl1.add(0.001, 0.001, 50);
-        RoadElevationInterpolator.smoothElevation(pl1);
+        GraphElevationSmoothing.smoothElevation(pl1);
         assertEquals(3, pl1.size());
         assertEquals(50, pl1.getElevation(1), .1);
 
@@ -44,7 +44,7 @@ public class RoadAverageElevationInterpolatorTest {
         pl2.add(0.0016, 0.0015, 150);
         pl2.add(0.0017, 0.0015, 220);
         pl2.add(0.002, 0.002, 20);
-        RoadElevationInterpolator.smoothElevation(pl2);
+        GraphElevationSmoothing.smoothElevation(pl2);
         assertEquals(5, pl2.size());
         assertEquals(120, pl2.getElevation(1), .1);
         // This is not 120 anymore, as the point at index 1 was smoothed from 160=>120
